requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.5.0
pymongo==4.6.1
python-dotenv==1.0.1
SQLAlchemy==2.0.25
